# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/samples/{id}/variants": {
            "get": {
                "description": "Get all variant files for a sample",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "variants"
                ],
                "summary": "Get sample variants",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Sample ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.VariantFile"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/variants": {
            "get": {
                "description": "Get all variant files",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "variants"
                ],
                "summary": "List variant files",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.VariantFile"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Add a new variant file record",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "variants"
                ],
                "summary": "Create variant file",
                "parameters": [
                    {
                        "description": "Variant file info",
                        "name": "variant_file",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.VariantFile"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.VariantFile"
                        }
                    }
                }
            }
        },
        "/api/variants/{id}": {
            "delete": {
                "description": "Delete variant file by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "variants"
                ],
                "summary": "Delete variant file",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Variant file ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.VariantFile": {
            "type": "object",
            "properties": {
                "checksum": {
                    "type": "string"
                },
                "file_path": {
                    "type": "string"
                },
                "file_type": {
                    "type": "string"
                },
                "genome_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "sample_id": {
                    "type": "integer"
                },
                "uploaded_at": {
                    "type": "string"
                },
                "uploaded_by": {
                    "type": "integer"
                }
            }
        }
    }
}